% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDist.R
\name{FDist}
\alias{FDist}
\title{Fit of univariate distributions with censored data ignored by default or can be inputed.}
\usage{
FDist(X, gen = 1, Cont = TRUE, inputNA, plot = FALSE,
  p.val_min = 0.05, criteria = 2, DPQR = T)
}
\arguments{
\item{X}{A numeric vector}

\item{gen}{A positive integer, indicates the samle length to be generated by the fit, 1 by default.}

\item{Cont}{TRUE, by defalut the distribution is considered as continuos, and belongs to fitdistplus distributions.}

\item{inputNA}{A number to replace censored values, if is missing, only non-censored values will be valuated.}

\item{plot}{FALSE, if TRUE, a plot showing the data distribution will be given.}

\item{p.val_min}{0.05, minimum p.value for Anderson Darling and KS Test to non-reject the null hypothesis and continue with the process.}

\item{criteria}{A positive integer, if 1, show the distributions wich were non-rejected by the Anderson Darling or Kolmogorov Smirnov tests, in other cases the criteria is that they mustn't be rejected by both.}

\item{DPQR}{TRUE, creates the distribution function, density and quantile function with the names dfit, pfit and qfit.}
}
\value{
Calculate the distribution name with params, a function to reproduce randm values from that distribution, a numeric vector of random numbers from that function, Anderson Darling and KS p.values, a plot showing the distribution difference between the real sample and the generated values and a list with the random deviates genetator, the distribution function, density and quantile function
}
\description{
Fit of univariate distributions with censored data ignored by default or can be inputed.
}
\examples{
FIT1<-FDist(rnorm(1000,10),p.val_min=.03,criteria=1,plot=TRUE)
#FIT1[[1]]
#FIT1[[2]]()
#FIT1[[4]]
#FIT1[[5]]



}
